added login on backend

+++ backend/rest_api/src/api/v1/schemas/goods.py
@@ -98,7 +98,7 @@ class GoodsInDB(GoodsInDBBase):
 
 class GoodsCountByName(BaseModel):
     name: str
-    count: int
+    count: Decimal
 
 
 class GoodsSummByName(BaseModel):
